25.6.8
test

+++ EngineCode/Mesh/Core/MeshType.h
@@ -21,8 +21,8 @@ struct FMeshRenderingData
 	vector<uint16_t> IndexData;
 public:
 	// 得到当前渲染对象顶点与索引size
-	UINT GetVertexSizeInBytes() { return VertexData.size() * sizeof(FVertex); }
-	UINT GetIndexSizeInBytes() { return IndexData.size() * sizeof(uint16_t); }
+	UINT GetVertexSizeInBytes() { return  (UINT)VertexData.size() * sizeof(FVertex); }
+	UINT GetIndexSizeInBytes() { return  (UINT)IndexData.size() * sizeof(uint16_t); }
 };
 
 enum EPyramidNumberSides
